worked on routes, models, and controllers - additional changes still needed
