+++ client/src/utils/API.js
@@ -3,7 +3,7 @@ import axios from "axios";
 export default {
   // Gets all books
   getBooks: function() {
-    return axios.get("/api/books");
+    return axios.get("/api/search");
   },
   // Gets the book with the given id
   getBook: function(id) {
