+++ models/index.js
@@ -1,3 +1,4 @@
 module.exports = {
-  Book: require("./book")
+  Book: require("./book"),
+  Search: require("./search")
 };
